Reorganize news folder structures
