+++ app/src/main/java/org/mozilla/rocket/content/news/domain/LoadNewsSettingsUseCase.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.mozilla.rocket.content.news
+package org.mozilla.rocket.content.news.domain
 
 import org.mozilla.rocket.content.MediatorUseCase
 import org.mozilla.rocket.content.Result
