+++ app/src/main/java/org/mozilla/rocket/content/news/ui/LanguageListDialogHelper.kt
@@ -1,4 +1,4 @@
-package org.mozilla.rocket.content.news
+package org.mozilla.rocket.content.news.ui
 
 import android.annotation.SuppressLint
 import android.content.Context
