+++ app/src/main/java/org/mozilla/rocket/content/news/ui/NewsActivity.kt
@@ -1,4 +1,4 @@
-package org.mozilla.rocket.content.news
+package org.mozilla.rocket.content.news.ui
 
 import android.content.Context
 import android.content.Intent
