+++ app/src/main/java/org/mozilla/rocket/content/news/ui/NewsCategoryPreference.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.rocket.content.news.data
+package org.mozilla.rocket.content.news.ui
 
 import android.content.Context
 import androidx.preference.Preference
@@ -16,6 +16,7 @@ import android.view.ViewGroup
 import android.widget.ProgressBar
 import android.widget.ToggleButton
 import org.mozilla.focus.R
+import org.mozilla.rocket.content.news.data.NewsCategory
 
 class NewsCategoryPreference @JvmOverloads constructor(context: Context, attributes: AttributeSet? = null) :
     Preference(context, attributes) {
