+++ app/src/main/java/org/mozilla/rocket/content/news/ui/NewsLanguagePreference.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.rocket.content.news.data
+package org.mozilla.rocket.content.news.ui
 
 import android.content.Context
 import androidx.preference.Preference
