+++ app/src/main/java/org/mozilla/rocket/content/news/ui/NewsSettingFragment.kt
@@ -1,4 +1,4 @@
-package org.mozilla.rocket.content.news
+package org.mozilla.rocket.content.news.ui
 
 import android.os.Bundle
 import android.util.Log
@@ -10,9 +10,7 @@ import org.mozilla.focus.R
 import org.mozilla.focus.telemetry.TelemetryWrapper
 import org.mozilla.rocket.content.appComponent
 import org.mozilla.rocket.content.news.data.NewsCategory
-import org.mozilla.rocket.content.news.data.NewsCategoryPreference
 import org.mozilla.rocket.content.news.data.NewsLanguage
-import org.mozilla.rocket.content.news.data.NewsLanguagePreference
 import org.mozilla.rocket.content.news.data.NewsSettingsRepository
 import javax.inject.Inject
